Beautification
Beautification is the process of making visual improvements to a person,
place, or thing. With regard to a town, city, or urban area, this most
often involves planting trees, shrubbery, and other greenery, but
frequently also includes adding decorative or historic-style street
lights and other lighting and replacing broken pavement, often with
brick or other natural materials. Old-fashioned cobblestones are
sometimes used for crosswalks; they provide the additional benefit of
slowing motorists.